success to show nav by using props with animation but need to adjust the position
