+++ src/server/views/widget/page_alerts.html
@@ -79,7 +79,6 @@
 
     {% if isTrashPage() %}
       <div id="trash-page-alert"></div>
-      <div id="custom-nav-bar"></div>
     {% endif %}
   </div>
 </div>
